refactored scrape to its own file in backend
